Remove old CMake workaround causing failure on Windows
The CMake for the samples use a work-around for linking the SYCL
libraries, however this causes failures on Windows due to explicit
linking rather than letting the DPC++ driver handle it correctly when
passed the -fsycl flag. This commit removes the workaround as it is
believed to no longer be needed.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>